+++ cakk-api/src/main/java/com/cakk/api/dto/param/link/ShopLinkParam.java
@@ -1,4 +1,4 @@
-package com.cakk.api.dto.request.link;
+package com.cakk.api.dto.param.link;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
