+++ cakk-api/src/main/java/com/cakk/api/service/shop/ShopService.java
@@ -43,9 +43,9 @@ import com.cakk.domain.mysql.entity.shop.CakeShopLink;
 import com.cakk.domain.mysql.entity.shop.CakeShopOperation;
 import com.cakk.domain.mysql.entity.user.BusinessInformation;
 import com.cakk.domain.mysql.entity.user.User;
-import com.cakk.domain.mysql.event.EventMapper;
 import com.cakk.domain.mysql.event.shop.CertificationEvent;
 import com.cakk.domain.mysql.event.views.CakeShopIncreaseViewsEvent;
+import com.cakk.domain.mysql.mapper.EventMapper;
 import com.cakk.domain.mysql.repository.reader.BusinessInformationReader;
 import com.cakk.domain.mysql.repository.reader.CakeShopReader;
 import com.cakk.domain.mysql.repository.reader.UserReader;
