+++ cakk-api/src/main/java/com/cakk/api/validator/OperationValidator.java
@@ -3,7 +3,6 @@ package com.cakk.api.validator;
 import static java.util.Objects.*;
 
 import java.util.EnumMap;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -11,7 +10,7 @@ import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 
 import com.cakk.api.annotation.OperationDay;
-import com.cakk.api.dto.request.operation.ShopOperationParam;
+import com.cakk.api.dto.param.operation.ShopOperationParam;
 import com.cakk.common.enums.Days;
 
 public class OperationValidator implements ConstraintValidator<OperationDay, List<ShopOperationParam>> {
