+++ cakk-domain/mysql/src/main/java/com/cakk/domain/mysql/entity/cake/Cake.java
@@ -26,7 +26,7 @@ import lombok.NoArgsConstructor;
 
 import com.cakk.domain.mysql.entity.audit.AuditEntity;
 import com.cakk.domain.mysql.entity.shop.CakeShop;
-import com.cakk.domain.mysql.event.EventMapper;
+import com.cakk.domain.mysql.mapper.EventMapper;
 import com.cakk.domain.mysql.event.views.CakeIncreaseViewsEvent;
 import com.cakk.domain.mysql.mapper.CakeTagMapper;
