+++ cakk-domain/mysql/src/main/java/com/cakk/domain/mysql/mapper/EventMapper.java
@@ -1,4 +1,4 @@
-package com.cakk.domain.mysql.event;
+package com.cakk.domain.mysql.mapper;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
